seventh Commit

+++ src/main/java/com/Atef/gestionstock/service/Impl/CategoryServiceImpl.java
@@ -3,6 +3,10 @@ package com.Atef.gestionstock.service.Impl;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import com.Atef.gestionstock.exception.InvalidOperationException;
+import com.Atef.gestionstock.model.Article;
+import com.Atef.gestionstock.model.LigneCommandeFournisseur;
+import com.Atef.gestionstock.repository.ArticleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
@@ -24,10 +28,12 @@ public class CategoryServiceImpl implements CategoryService{
 
 	
 	private CategoryRepository categoryRepository;
-	
+	private ArticleRepository articleRepository;
+
 	@Autowired
-	public CategoryServiceImpl(CategoryRepository categoryRepository) {
+	public CategoryServiceImpl(CategoryRepository categoryRepository, ArticleRepository articleRepository) {
 		this.categoryRepository = categoryRepository;
+		this.articleRepository = articleRepository;
 	}
 	
 	
@@ -84,6 +90,10 @@ public class CategoryServiceImpl implements CategoryService{
 		if (id ==null) {
 			log.error("Category id is null");
 			return ;
+		}
+		List<Article> articles =articleRepository.findAllByCategoryId(id);
+		if (!articles.isEmpty()){
+			throw new InvalidOperationException("Impossible de supprimer cette categorie qui est déja utilisé ",ErrorCodes.CATEGORY_ALREADY_IN_USE);
 		}
 		categoryRepository.deleteById(id);
